test(cli): Add unit tests for utils pkg
Added test table for RunCloc method in utils.go. Have not yet added
tests to review the contents of the output file but covered basic error
reporting.

Could not mock a git repo to use when testing the opitional gitObj arg
but I think I'll just ignore that aspect in my testing or tackle it in
the future.