refactor(cli): Remove temp file for cloc output
As we need to clean the output provided by cloc to work with our JS
charting libraries, it makes the most sense to send the STDOUT of cloc
to another part of our program for parsing and cleaning.

The usage of a temporary file creates unncessary resources (albeit
temporarily) so I've taken this approach to keep the file system
untouched.

Tests for RunCloc have been updated; note that we choose to ignore the
output of cloc since we assume that result to be well tested and
accurate. Have commented out other functions in utils.go and will remove
if deemed irrelevant.