style(cli): Rename utils.go to data.go
Changed utils.go naming as it matched the package name but did not
provide any clarity as to what it entailed. The methods noted in what is
now know as data revolve around data extraction, munging, and formatting
into JSON. As such, the name data.go is more indicative of the file's
contents and has been changed accordingly.