feat(cli): Implement creation of nested ChartObj
The Update method in chart.go essentially uses its receiver instance as the
root of a tree and takes in parameters to create new ChartObjs
and grow the branches of this tree/graph.

The algorithm identifies whether or not a child dir is present and
recursively creates children nodes as deemed necessary. Tests in
chart_test.go should cover primary use-cases.

Currently used to implement GetLinesByFile in data.go but a similar
approach should be leveraged for GetLinesByLang.