feat(cli): Support language to color mappings
To ensure that languages are displayed with their appropriate color in
the visualization aspect of the project, I've included a map of language
name to hex code (representing the language's official color).

Languages not covered in the scope of this map should be assigned to a
random hex value. Tests are added to cover both cases.