refactor(cli): Convert ChartObj to interface
Since a ChartObj can represent either a directory or file node in our
filesystem structure, it made the most sense to convert ChartObj into an
interface and have Directory and File classes implement it.

Our root is a Directory, which contains a slice of ChartObjs to
represent children. This relationship is recursive in nature and
continues down the tree.

Updated methods that utilize ChartObjs and fixed tests to ensure that
object state is as expected after multiple calls to Update.