feat(cli): Update behavior of random color
In the case we try and find a color not noted in our COLORS map, our
program determines a random color and assigns it the current node. The
issue with the current implementation is the fact that subsequent nodes
that match the same language will undergo the same randomization; thus,
we have an issue where files of the same language have different colors.

To remedy this, we simply add the generated color from our
randomization back to the COLORS map for the duration of the program's
execution.