feat(visuals): Embed chart into HTML template
Utilized Go's html/templates libary to insert Go structs and variables
into an HTML template. CSS and general stylistic updates need to be made
but the MVP is working as intended.

Created a separate .gohtml template file for this purpose; unsure
whether or not it gets compiled with the rest of the files so something
to keep an eye on.