feat(template): Implement opening in browser
Before this fix, the program generated out.html but required the user to
manually open it in their desired browser.

This change utilizes allows the output of the program to be
automatically opened in the user's default browser. Determination of
browser is OS specific but should work for all major OSs.