feat(utils): Add support for remote Git repos
Adds functionality to run cloc and clocviz on remote repositories. Upon
the user passing an argument formatted as so: "user/repoName", the
program will perform an analysis on the target repo.

Note that this method requires the cloning of the remote repo and thus
is quite a slow process. Could not determine a more efficient way to do
this and since it seems to give accurate results, leaving for now.

Still open to add support for git objects.