refactor(cloc): Wrap cloc helpers into one method
Since there are two primary applications of running cloc (local
directory or git repo), it was deemed appropriate to wrap them in a
generic function.

I took the old RunCloc method and made that wrap around
runClocOnLocalDir and runClocOnGitRepo.

Refactored tests to ensure that it still works as expected and updated
main to produce valid output.