fix(data): Removed git object from RunCloc
The cloc documentation states that it works on git objects (commit
hashes, branches, etc) but from playing around with the tool, it doesn't
appear to be working as intended.

Instead of trying to troubleshoot, I decided to simplify the program and
have it only work on the current snapshot of the repository. Perhaps
there is room to update later on but for now, this seems to be cleanest.

Updated any methods that use a "gitObj" and ensured that tests reflected
these changes.