fix(chart): Implement error checking for Update()
The Update method in chart.go references the current part of the path by
indexing path[0]. As such, in the case that the path is empty, an
IndexOutOfBounds exception is triggered.

To remedy this, I've added a simple check to determine the size of the
path and terminate if the path is empty, preventing the program from
ever reaching the path[0] assignment.