refactor(color): Convert hex to RGB struct
As our color gradation algorithm will require constant modification of
the hex values associated with each chart node, I thought it would be
best to convert it to RGB for easier manipulation.

With the color attribute of Directory and File nodes now being of type
RGB, we can easily access the red, green, and blue attributes and update
them easily.

Changed all instances where color was a string and converted over to RGB
structs.

The behavior of JSON creation is a bit wonky but it looks like
the correct data is being fed to the front-end. The program after this
commit includes those JSON methods but does not actually invoke them.
Keeping in just in case they are needed later.