feat(visuals): Merge Fabrizzio's fork - 0c5c930c5e
Wanted to see how the backend and frontend came together after
Fabrizzio's updates. Had to make minor modifications to ensure that the
ChartObj struct was properly serialized into JSON and rendered in our
template.

The legend breaks if we use ByFile as our data. Need to initialize
that part using the ByLang JSON. Should be remedied once we include
both JSON objects charts.