fix(static): Fix rendering of CSS and JS in index
The compiled binary was having issues working outside the clocviz
directory; the HTML was rendering but the CSS and JS (from styles.css
and script.js) were not working as expected.

Updated rice box through "rice embed-go" (need to add hook to do this
automatically) and updated index.tmpl links and sytlesheets from
"src/static/..." to "/src/static/...".

Works now but let's see if it persists.